Change require message to claimer-proxy-already-exists.
The previous message proxy-registered-to-owner was a bit confusing, as
it is not clear if proxy is the new proxy or the old proxy.

Also owner was a bit ambiguous since it is not guaranteed that owner parameter
is the msg.sender (there is a check for that beforehand but someone just
seeing the require message might not neccesarily know that).